pyproject.toml
[project]
name = "youtube-agent"
version = "0.1.0"
description = "YouTube自動化AIエージェントシステム"
requires-python = ">=3.11"

dependencies = [
    # AI / LLM
    "anthropic>=0.34.0",

    # RAG
    "llama-index>=0.10.0",
    "llama-index-readers-notion>=0.1.0",
    "llama-index-embeddings-openai>=0.1.0",

    # HTTP / API
    "httpx>=0.27.0",
    "google-api-python-client>=2.140.0",
    "google-auth-oauthlib>=1.2.0",
    "notion-client>=2.2.0",

    # TTS
    "elevenlabs>=1.3.0",
    "openai>=1.40.0",

    # 動画処理
    "moviepy>=1.0.3",
    "Pillow>=10.4.0",

    # CLI
    "typer>=0.12.0",
    "rich>=13.7.0",

    # 設定管理
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.5.0",
    "mypy>=1.11.0",
]

[project.scripts]
youtube-agent = "cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "cli", "config", "models", "pipelines", "rag", "tts", "video", "youtube"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.11"
strict = true
